Replaced inner class Account with outer class
